engineers_core/requirements.txt
beautifulsoup4==4.7.1
bottlenose==1.1.8
bs4==0.0.1
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==2.1
django-allauth==0.39.1
django-cors-headers==2.5.3
django-environ==0.4.5
django-filter==2.1.0
django-rest-auth==0.9.5
django-rest-swagger==2.0.7
django-templated-mail==1.1.1
djangorestframework==3.9.2
djangorestframework-bulk==0.2.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
djoser==1.5.1
drf-nested-routers==0.91
feedparser==5.2.1
get==2019.4.13
google-api-core==1.13.0
google-auth==1.6.3
google-cloud-core==1.0.2
google-cloud-storage==1.16.1
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
gspread==3.1.0
gunicorn==19.9.0
httplib2==0.12.3
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
lxml==4.3.3
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
numpy==1.16.4
oauth2client==4.1.3
oauthlib==3.0.1
openapi-codec==1.3.2
opencv-contrib-python==4.1.0.25
Pillow==6.0.0
post==2019.4.13
protobuf==3.8.0
public==2019.4.13
pyasn1==0.4.5
pyasn1-modules==0.2.5
PyJWT==1.7.1
PyMySQL==0.9.3
PySocks==1.6.8
python3-openid==3.1.0
pytz==2019.1
query-string==2019.4.13
request==2019.4.13
requests==2.21.0
requests-oauthlib==1.2.0
responses==0.10.6
rsa==4.0
selenium==3.141.0
simplejson==3.16.0
six==1.9.0
soupsieve==1.9.1
SQLAlchemy==1.3.3
sqlparse==0.3.0
tweepy==3.7.0
uritemplate==3.0.0
urllib3==1.24.2
